LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.4_String_Literals/S7.8.4_A6.2_T2-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: \x can only be followed by a hex character sequence
S7.8.4_A6.2_T2

PASS Expected parsing failure

TEST COMPLETE
